feat: create skeleton list common component
