--- app/_components/features/main/PeriodList/SkeletonList.module.scss
@@ -1,20 +1,0 @@
-.skeleton-list {
-  width: 100%;
-  display: flex;
-  gap: 30px;
-  overflow-x: hidden;
-  padding-left: 30px;
-
-  .skeleton-wrapper {
-    box-sizing: border-box;
-    width: 320px;
-    display: flex;
-    gap: 10px;
-
-    .infos {
-      display: flex;
-      flex-direction: column;
-      gap: 20px;
-    }
-  }
-}
